Fix credits error
I am one of the OpenBVE translators and I noticed that the credits here had a small mistake (I was included in "Packagers" instead of "Translators". Also, both my email and website have changed (and may change again in the future), so I think it is better to have only my name included to make this future-proof.

Thank you very much!